fix: status 클래스 외부로 분리
