+++ src/main/java/com/airdnb/image/entity/ImageStatus.java
@@ -1,0 +1,5 @@
+package com.airdnb.image.entity;
+
+public enum ImageStatus {
+    ACTIVE, DELETED
+}
