+++ src/main/java/com/airdnb/reservation/ReservationService.java
@@ -7,8 +7,8 @@ import com.airdnb.member.entity.Member;
 import com.airdnb.reservation.dto.ReservationCreateRequest;
 import com.airdnb.reservation.dto.ReservationQueryResponse;
 import com.airdnb.reservation.entity.Reservation;
-import com.airdnb.reservation.entity.Reservation.ReservationStatus;
 import com.airdnb.reservation.entity.ReservationPeriod;
+import com.airdnb.reservation.entity.ReservationStatus;
 import com.airdnb.stay.StayService;
 import com.airdnb.stay.entity.Stay;
 import java.time.LocalDate;
