+++ src/main/java/com/airdnb/reservation/entity/Reservation.java
@@ -2,7 +2,6 @@ package com.airdnb.reservation.entity;
 
 import com.airdnb.global.exception.ForbiddenException;
 import com.airdnb.global.exception.InvalidRequestException;
-import com.airdnb.global.exception.NotFoundException;
 import com.airdnb.member.entity.Member;
 import com.airdnb.stay.entity.Stay;
 import jakarta.persistence.Embedded;
@@ -16,7 +15,6 @@ import jakarta.persistence.Id;
 import jakarta.persistence.JoinColumn;
 import jakarta.persistence.ManyToOne;
 import java.time.LocalDateTime;
-import java.util.Arrays;
 import lombok.AccessLevel;
 import lombok.Builder;
 import lombok.Getter;
@@ -87,17 +85,6 @@ public class Reservation {
     public void validateQueryAuthority(String currentMemberId) {
         if (!isCustomer(currentMemberId) && !isHost(currentMemberId)) {
             throw new ForbiddenException("예약 조회 권한이 없는 사용자입니다.");
-        }
-    }
-
-    public enum ReservationStatus {
-        PENDING, APPROVED, REJECTED, CANCELED;
-
-        public static ReservationStatus of(String statusValue) {
-            return Arrays.stream(values())
-                    .filter(reservationStatus -> reservationStatus.name().equals(statusValue.toUpperCase()))
-                    .findAny()
-                    .orElseThrow(() -> new NotFoundException("일치하는 예약 상태를 찾을 수 없습니다."));
         }
     }
 }
